+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## dev
 
-## Beta (0.1.5)
+## Beta (0.1.6)
 
 - Initial release of the library
 - Fixed missing data package
 - Fixed Biperby
 - Add missing en model
 - Fix error in SpaCy model
+- Improve code and fix `get_num_sentences` and `nws_90`